Rename AreBooking into AreaStatus and vice versa
The purposes of these two views are flipped with these changes.
