+++ app/components/AreaBooking/messages.js
@@ -1,20 +1,21 @@
+/*
+ * HomePage Messages
+ *
+ * This contains all the text for the HomePage component.
+ */
 import { defineMessages } from 'react-intl';
 
 export default defineMessages({
-  submitButton: {
-    id: 'boilerplate.components.AreaBooking.button.submit',
-    defaultMessage: 'Varaa',
+  resourceIsAvailable: {
+    id: 'AreaStatus.area.resourceIsAvailable',
+    defaultMessage: 'Tila on vapaa',
   },
-  cancel: {
-    id: 'boilerplate.components.Confirm.cancel',
-    defaultMessage: 'Peruuta varaus',
+  resourceIsNotAvailable: {
+    id: 'AreaStatus.area.resourceIsNotAvailable',
+    defaultMessage: 'Tila on varattu',
   },
-  reset: {
-    id: 'boilerplate.components.AreaBooking.link.reset',
-    defaultMessage: 'Palaa alkuun',
-  },
-  finishedLink: {
-    id: 'boilerplate.components.Confirm.finishedLink',
-    defaultMessage: 'Valmis!',
+  showMore: {
+    id: 'AreaStatus.ShowMoreButton.text',
+    defaultMessage: 'Tilan tarkemmat tiedot',
   },
 });
