+++ app/components/AreaStatus/Wrapper.js
@@ -1,18 +1,21 @@
 import styled from 'styled-components';
 
-const Wrapper = styled.div`
+export const Wrapper = styled.div`
   position: relative;
   display: flex;
-  flex-direction: column;
-  justify-content: center;
+  align-items: center;
+  justify-items: center;
   width: 50%;
+  flex-grow: 1;
+  background-color: ${props => props.theme.secondaryColor};
   padding: 5%;
-  background-color: ${props => props.theme.primaryColor};
-
-  &.hide-on-toggle,
-  .hide-on-toggle {
-    display: none;
-  }
+  background-image: url(${props => props.theme.bgImage});
+  background-repeat: repeat-y;
+  background-size: auto 100%;
+  background-position: 0 0;
 `;
 
-export default Wrapper;
+export const Div = styled.div`
+  width: 100%;
+  text-align: center;
+`;
