+++ app/components/AreaStatus/areaStatusStyles.js
@@ -1,9 +1,9 @@
 import { css } from 'styled-components';
 
-const areaBookingStyles = css`
+const areaStatusStyles = css`
   display: flex;
   align-items: stretch;
   height: 100%;
 `;
 
-export default areaBookingStyles;
+export default areaStatusStyles;
