+++ app/components/AreaStatus/bg-wave-green.png
