+++ app/components/AreaStatus/index.js
@@ -1,80 +1,152 @@
-/* eslint-disable react/prop-types */
-import React from 'react';
-import Upcoming from 'components/Upcoming';
-import Clock from 'components/Clock';
-import Status from 'components/Status';
-import SlideUpContent from 'components/SlideUpContent';
+import React, { useState, useLayoutEffect, useRef } from 'react';
+import LocaleToggle from 'containers/LocaleToggle';
+import PropTypes from 'prop-types';
+import camelCaseKeys from 'camelcase-keys';
+
+import { createStructuredSelector } from 'reselect';
 import { connect } from 'react-redux';
 import { compose } from 'redux';
-import { createStructuredSelector } from 'reselect';
+import injectReducer from 'utils/injectReducer';
+import SceneCancel from 'components/SceneCancel';
+import SceneAction from 'components/SceneAction';
+import SceneVerify from 'components/SceneVerify';
+import SceneError from 'components/SceneError';
+import SceneSetup from 'components/SceneSetup';
+import SceneLoading from 'components/SceneLoading';
+import SceneStrongAuth from 'components/SceneStrongAuth';
+import reducer from 'containers/HomePage/reducer';
+
 import {
-  makeSelectDate,
-  makeSelectNextAvailableTime,
-  makeSelectAvailableUntil,
-  makeSelectIsDescriptionOpen,
+  makeSelectScene,
+  makeSelectSelectedSlot,
+  makeSelectErrorMessage,
+  makeSelectResourceId,
+  makeSelectResource,
 } from 'containers/HomePage/selectors';
-import { toggleIsDescriptionOpen } from 'containers/HomePage/actions';
+import { changeScene, makeReservation } from 'containers/HomePage/actions';
+import Calendar from 'components/Calendar';
+import { Wrapper, Div } from './Wrapper';
 
-import Wrapper from './Wrapper';
+function useElementSize(ref) {
+  const [size, setSize] = useState([0, 0]);
 
-/* eslint-disable react/prefer-stateless-function */
-class AreaStatus extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {};
-  }
+  useLayoutEffect(() => {
+    function updateSize() {
+      const currentElement = ref.current;
+      if (currentElement) {
+        setSize([ref.current.clientWidth, ref.current.clientHeight]);
+      }
+    }
 
-  render() {
-    const wrapperClass = Object.entries({
-      'slide-down': true,
-      'hide-on-toggle': this.props.isCondensed,
-    })
-      .filter(([, isIncluded]) => isIncluded)
-      .map(([className]) => className)
-      .join(' ');
+    window.addEventListener('resize', updateSize);
+    updateSize();
 
-    return (
-      <Wrapper className={wrapperClass}>
-        <Clock
-          className={this.props.isDescriptionOpen}
-          date={this.props.date}
-        />
+    return () => {
+      window.removeEventListener('resize', updateSize);
+    };
+  }, []);
 
-        <Status
-          resourceName={this.props.resourceName}
-          nextAvailableTime={this.props.nextAvailableTime}
-          availableUntil={this.props.availableUntil}
-          isResourceAvailable={this.props.isResourceAvailable}
-        />
+  return size;
+}
 
-        <Upcoming
-          className={this.state.isHidden}
-          upcomingReservations={this.props.upcomingReservations}
-        />
+const AreaStatus = ({
+  resource,
+  scene,
+  selectedSlot,
+  onChangeSceneToStart,
+  onMakeReservation,
+  onChangeSceneToCancel,
+  onCalendarViewChange,
+  errorMessage,
+  resourceId,
+}) => {
+  const wrapperRef = useRef(null);
+  const [, height] = useElementSize(wrapperRef);
 
-        {this.props.resourceDescription && (
-          <SlideUpContent
-            visible={this.props.isDescriptionOpen}
-            content={this.props.resourceDescription}
-            onButtonClick={() => this.props.onToggleDescriptionOpen()}
+  return (
+    <Wrapper innerRef={wrapperRef}>
+      <Div>
+        <LocaleToggle />
+        {resource &&
+          height && (
+            <Calendar
+              // Approximately remove padding from wrapper height
+              height={height - 2 * 84}
+              // Transform immutable data structure into a JSON object,
+              // and transform snake_case in that object into camelCase.
+              resource={camelCaseKeys(resource.toJS())}
+              onDateChange={() => {}}
+              onTimeChange={() => {}}
+              onViewTypeChange={onCalendarViewChange}
+            />
+          )}
+
+        {scene === 'Loading' && <SceneLoading />}
+        {scene === 'Setup' && <SceneSetup />}
+        {scene === 'Action' && (
+          <SceneAction
+            onTimesUp={onChangeSceneToStart}
+            selectedSlot={selectedSlot}
+            onButtonClick={onMakeReservation}
+            onCancelClick={onChangeSceneToCancel}
           />
         )}
-      </Wrapper>
-    );
-  }
-}
+        {scene === 'Cancel' && (
+          <SceneCancel
+            onTimesUp={onChangeSceneToStart}
+            onButtonClick={onChangeSceneToStart}
+          />
+        )}
+        {scene === 'Verify' && (
+          <SceneVerify
+            onTimesUp={onChangeSceneToStart}
+            onButtonClick={onChangeSceneToStart}
+          />
+        )}
+        {scene === 'Error' && (
+          <SceneError
+            errorMessage={errorMessage}
+            onButtonClick={onChangeSceneToStart}
+          />
+        )}
+        {scene === 'StrongAuth' && (
+          <SceneStrongAuth
+            resource={resource}
+            resourceId={resourceId}
+            errorMessage={errorMessage}
+          />
+        )}
+      </Div>
+    </Wrapper>
+  );
+};
+
+AreaStatus.propTypes = {
+  scene: PropTypes.any,
+  selectedSlot: PropTypes.any,
+  onChangeSceneToStart: PropTypes.any,
+  onChangeSceneToCancel: PropTypes.any,
+  onMakeReservation: PropTypes.any,
+  onCalendarViewChange: PropTypes.func,
+  errorMessage: PropTypes.any,
+  resource: PropTypes.any,
+  resourceId: PropTypes.any,
+};
 
 export function mapDispatchToProps(dispatch) {
   return {
-    onToggleDescriptionOpen: () => dispatch(toggleIsDescriptionOpen()),
+    onMakeReservation: () => dispatch(makeReservation()),
+    onChangeSceneToStart: () => dispatch(changeScene('Start')),
+    onChangeSceneToCancel: () => dispatch(changeScene('Cancel')),
   };
 }
 
 const mapStateToProps = createStructuredSelector({
-  date: makeSelectDate(),
-  nextAvailableTime: makeSelectNextAvailableTime(),
-  availableUntil: makeSelectAvailableUntil(),
-  isDescriptionOpen: makeSelectIsDescriptionOpen(),
+  scene: makeSelectScene(),
+  selectedSlot: makeSelectSelectedSlot(),
+  errorMessage: makeSelectErrorMessage(),
+  resourceId: makeSelectResourceId(),
+  resource: makeSelectResource(),
 });
 
 const withConnect = connect(
@@ -82,4 +154,9 @@ const withConnect = connect(
   mapDispatchToProps,
 );
 
-export default compose(withConnect)(AreaStatus);
+const withReducer = injectReducer({ key: 'home', reducer });
+
+export default compose(
+  withReducer,
+  withConnect,
+)(AreaStatus);
